LPS-151456 Set portlet namespace as default `namespace`
